+++ packages/language-service/test/diagnostics_spec.ts
@@ -26,7 +26,6 @@ describe('diagnostics', () => {
     const service = ts.createLanguageService(mockHost, documentRegistry);
     ngHost = new TypeScriptServiceHost(mockHost, service);
     ngService = createLanguageService(ngHost);
-    ngHost.setSite(ngService);
   });
 
   it('should be no diagnostics for test.ng',
