+++ packages/language-service/test/hover_spec.ts
@@ -23,7 +23,6 @@ describe('hover', () => {
   let service = ts.createLanguageService(mockHost, documentRegistry);
   let ngHost = new TypeScriptServiceHost(mockHost, service);
   let ngService = createLanguageService(ngHost);
-  ngHost.setSite(ngService);
 
 
   it('should be able to find field in an interpolation', () => {
